Move commonly used E2E methods into a utils file
Created separate utils file for methods only used in E2E test, as opposed
to the testutils which is used by integration and E2E tests.
