+++ e2e/condition_catalog_service_test.go
@@ -4,15 +4,12 @@
 package e2e
 
 import (
-	"encoding/json"
 	"fmt"
-	"net/http"
 	"path/filepath"
 	"testing"
 	"time"
 
 	"github.com/hashicorp/consul-terraform-sync/api"
-	"github.com/hashicorp/consul-terraform-sync/event"
 	"github.com/hashicorp/consul-terraform-sync/templates/tftmpl"
 	"github.com/hashicorp/consul-terraform-sync/testutils"
 	"github.com/hashicorp/consul/sdk/testutil"
@@ -532,30 +529,4 @@ func testCatalogServicesNoTagsTrigger(t *testing.T, taskConf, taskName, tempDirN
 	assert.Contains(t, content, "tag_c")
 
 	cleanup()
-}
-
-// eventCount returns number of events that are stored for a given task by
-// querying the Task Status API. Note: events have a storage limit (currently 5)
-func eventCount(t *testing.T, taskName string, port int) int {
-	events := events(t, taskName, port)
-	return len(events)
-}
-
-// events returns the events that are stored for a given task by querying the
-// Task Status API. Note: events have a storage limit (currently 5)
-func events(t *testing.T, taskName string, port int) []event.Event {
-	u := fmt.Sprintf("http://localhost:%d/%s/status/tasks/%s?include=events",
-		port, "v1", taskName)
-	resp := testutils.RequestHTTP(t, http.MethodGet, u, "")
-	defer resp.Body.Close()
-	assert.Equal(t, http.StatusOK, resp.StatusCode)
-
-	var taskStatuses map[string]api.TaskStatus
-	decoder := json.NewDecoder(resp.Body)
-	err := decoder.Decode(&taskStatuses)
-	require.NoError(t, err)
-
-	taskStatus, ok := taskStatuses[taskName]
-	require.True(t, ok, taskStatuses)
-	return taskStatus.Events
 }
